refactor to make execute a completely decoupled module
