+++ server/src-lib/Hasura/GraphQL/Execute.hs
@@ -1,10 +1,11 @@
 module Hasura.GraphQL.Execute
   ( GQExecPlan(..)
-  , ExecPlanResolved
-  , ResolvedOp(..)
-  , ExecPlanUnresolved
 
-  , getExecPlan
+  , ExecPlanPartial
+  , getExecPlanPartial
+
+  , ExecOp(..)
+  , ExecPlanResolved
   , getResolvedExecPlan
   , execRemoteGQ
 
@@ -12,13 +13,11 @@ module Hasura.GraphQL.Execute
   , EP.initPlanCache
   , EP.clearPlanCache
   , EP.dumpPlanCache
-
-  , module EL
-
   ) where
 
 import           Control.Exception                      (try)
 import           Control.Lens
+import           Data.Has
 
 import qualified Data.ByteString.Lazy                   as BL
 import qualified Data.CaseInsensitive                   as CI
@@ -32,42 +31,78 @@ import qualified Network.HTTP.Types                     as N
 import qualified Network.Wreq                           as Wreq
 
 import           Hasura.EncJSON
+import           Hasura.GraphQL.Context
+import           Hasura.GraphQL.Resolve.Context
 import           Hasura.GraphQL.Schema
 import           Hasura.GraphQL.Transport.HTTP.Protocol
+import           Hasura.GraphQL.Validate.Types
 import           Hasura.HTTP
 import           Hasura.Prelude
 import           Hasura.RQL.DDL.Headers
 import           Hasura.RQL.Types
 
-import qualified Hasura.GraphQL.Execute.Context         as EC
-import           Hasura.GraphQL.Execute.LiveQuery       as EL
+import qualified Hasura.GraphQL.Execute.LiveQuery       as EL
 import qualified Hasura.GraphQL.Execute.Plan            as EP
+import qualified Hasura.GraphQL.Execute.Query           as EQ
+
+import qualified Hasura.GraphQL.Resolve                 as GR
 import qualified Hasura.GraphQL.Validate                as VQ
 import qualified Hasura.GraphQL.Validate.Types          as VT
 
+-- The current execution plan of a graphql operation, it is
+-- currently, either local pg execution or a remote execution
+--
+-- The 'a' is parameterised so this AST can represent
+-- intermediate passes
 data GQExecPlan a
   = GExPHasura !a
   | GExPRemote !RemoteSchemaInfo !G.TypedOperationDefinition
   deriving (Functor, Foldable, Traversable)
 
-type ExecPlanUnresolved
+-- Enforces the current limitation
+assertSameLocationNodes
+  :: (MonadError QErr m) => [VT.TypeLoc] -> m VT.TypeLoc
+assertSameLocationNodes typeLocs =
+  case Set.toList (Set.fromList typeLocs) of
+    -- this shouldn't happen
+    []    -> return VT.HasuraType
+    [loc] -> return loc
+    _     -> throw400 NotSupported msg
+  where
+    msg = "cannot mix top level fields from two different graphql servers"
+
+-- TODO: we should fix this function asap
+-- as this will fail when there is a fragment at the top level
+getTopLevelNodes :: G.TypedOperationDefinition -> [G.Name]
+getTopLevelNodes opDef =
+  mapMaybe f $ G._todSelectionSet opDef
+  where
+    f = \case
+      G.SelectionField fld        -> Just $ G._fName fld
+      G.SelectionFragmentSpread _ -> Nothing
+      G.SelectionInlineFragment _ -> Nothing
+
+gatherTypeLocs :: GCtx -> [G.Name] -> [VT.TypeLoc]
+gatherTypeLocs gCtx nodes =
+  catMaybes $ flip map nodes $ \node ->
+    VT._fiLoc <$> Map.lookup node schemaNodes
+  where
+    schemaNodes =
+      let qr = VT._otiFields $ _gQueryRoot gCtx
+          mr = VT._otiFields <$> _gMutRoot gCtx
+      in maybe qr (Map.union qr) mr
+
+-- This is for when the graphql query is validated
+type ExecPlanPartial
   = GQExecPlan (GCtx, VQ.RootSelSet, [G.VariableDefinition])
 
-data ResolvedOp
-  = ROQuery !LazyRespTx
-  | ROMutation !LazyRespTx
-  | ROSubs !EL.LiveQueryOp
-
-type ExecPlanResolved
-  = GQExecPlan ResolvedOp
-
-getExecPlan
+getExecPlanPartial
   :: (MonadError QErr m)
   => UserInfo
   -> SchemaCache
   -> GQLReqParsed
-  -> m ExecPlanUnresolved
-getExecPlan userInfo sc req = do
+  -> m ExecPlanPartial
+getExecPlanPartial userInfo sc req = do
 
   (gCtx, _)  <- flip runStateT sc $ getGCtx (userRole userInfo) gCtxRoleMap
   queryParts <- flip runReaderT gCtx $ VQ.getQueryParts req
@@ -90,6 +125,18 @@ getExecPlan userInfo sc req = do
   where
     gCtxRoleMap = scGCtxMap sc
 
+-- An execution operation, in case of
+-- queries and mutations it is just a transaction
+-- to be executed
+data ExecOp
+  = ExOpQuery !LazyRespTx
+  | ExOpMutation !LazyRespTx
+  | ExOpSubs !EL.LiveQueryOp
+
+-- The graphql query is resolved into an execution operation
+type ExecPlanResolved
+  = GQExecPlan ExecOp
+
 getResolvedExecPlan
   :: (MonadError QErr m, MonadIO m)
   => EP.PlanCache
@@ -101,37 +148,152 @@ getResolvedExecPlan
   -> m ExecPlanResolved
 getResolvedExecPlan planCache userInfo sqlGenCtx sc scVer reqUnparsed = do
   planM <- liftIO $ EP.getPlan scVer (userRole userInfo)
-                opNameM queryStr planCache
+           opNameM queryStr planCache
   case planM of
     -- plans are only for queries and subscriptions
-    Just plan ->
-      GExPHasura <$> case plan of
-        EP.RPQuery queryPlan -> ROQuery <$> EP.mkNewQueryTx queryVars queryPlan
-        EP.RPSubs subsPlan -> ROSubs <$> EP.mkSubsOp queryVars subsPlan
+    Just plan -> GExPHasura <$> case plan of
+      EP.RPQuery queryPlan ->
+        ExOpQuery <$> EQ.queryOpFromPlan queryVars queryPlan
+      EP.RPSubs subsPlan ->
+        ExOpSubs <$> EL.subsOpFromPlan queryVars subsPlan
     Nothing -> noExistingPlan
   where
     GQLReq opNameM queryStr queryVars = reqUnparsed
+    addPlanToCache plan =
+      liftIO $ EP.addPlan scVer (userRole userInfo)
+      opNameM queryStr plan planCache
     noExistingPlan = do
       req      <- toParsed reqUnparsed
-      execPlan <- getExecPlan userInfo sc req
-      forM execPlan $ \(gCtx, rootSelSet, varDefs) ->
+      partialExecPlan <- getExecPlanPartial userInfo sc req
+      forM partialExecPlan $ \(gCtx, rootSelSet, varDefs) ->
         case rootSelSet of
           VQ.RMutation selSet ->
-            ROMutation <$> EC.getMutTx gCtx sqlGenCtx userInfo selSet
+            ExOpMutation <$> getMutOp gCtx sqlGenCtx userInfo selSet
           VQ.RQuery selSet -> do
-            (queryTx, planM) <- EC.getQueryTx gCtx sqlGenCtx
+            (queryTx, planM) <- getQueryOp gCtx sqlGenCtx
                                 userInfo selSet varDefs
-            forM_ planM $ \plan ->
-              liftIO $ EP.addPlan scVer (userRole userInfo)
-              opNameM queryStr (EP.RPQuery plan) planCache
-            return $ ROQuery queryTx
+            mapM_ (addPlanToCache . EP.RPQuery) planM
+            return $ ExOpQuery queryTx
           VQ.RSubscription fld -> do
-            (lqOp, planM) <- EC.getSubsOp gCtx sqlGenCtx
+            (lqOp, planM) <- getSubsOp gCtx sqlGenCtx
                              userInfo fld varDefs
-            forM_ planM $ \plan ->
-              liftIO $ EP.addPlan scVer (userRole userInfo)
-              opNameM queryStr (EP.RPSubs plan) planCache
-            return $ ROSubs lqOp
+            mapM_ (addPlanToCache . EP.RPSubs) planM
+            return $ ExOpSubs lqOp
+
+-- Monad for resolving a hasura query/mutation
+type E =
+  ReaderT ( UserInfo
+          , OpCtxMap
+          , TypeMap
+          , FieldMap
+          , OrdByCtx
+          , InsCtxMap
+          , SQLGenCtx
+          ) (Except QErr)
+
+runE
+  :: (MonadError QErr m)
+  => GCtx
+  -> SQLGenCtx
+  -> UserInfo
+  -> E a
+  -> m a
+runE ctx sqlGenCtx userInfo action =
+  either throwError return $
+  runExcept $ runReaderT action
+  (userInfo, opCtxMap, typeMap, fldMap, ordByCtx, insCtxMap, sqlGenCtx)
+  where
+    opCtxMap = _gOpCtxMap ctx
+    typeMap = _gTypes ctx
+    fldMap = _gFields ctx
+    ordByCtx = _gOrdByCtx ctx
+    insCtxMap = _gInsCtxMap ctx
+
+getQueryOp
+  :: (MonadError QErr m)
+  => GCtx
+  -> SQLGenCtx
+  -> UserInfo
+  -> VQ.SelSet
+  -> [G.VariableDefinition]
+  -> m (LazyRespTx, Maybe EQ.ReusableQueryPlan)
+getQueryOp gCtx sqlGenCtx userInfo fields varDefs =
+  runE gCtx sqlGenCtx userInfo $ EQ.convertQuerySelSet varDefs fields
+
+mutationRootName :: Text
+mutationRootName = "mutation_root"
+
+resolveMutSelSet
+  :: ( MonadError QErr m
+     , MonadReader r m
+     , Has UserInfo r
+     , Has OpCtxMap r
+     , Has FieldMap r
+     , Has OrdByCtx r
+     , Has SQLGenCtx r
+     , Has InsCtxMap r
+     )
+  => VQ.SelSet
+  -> m LazyRespTx
+resolveMutSelSet fields = do
+  aliasedTxs <- forM (toList fields) $ \fld -> do
+    fldRespTx <- case VQ._fName fld of
+      "__typename" -> return $ return $ encJFromJValue mutationRootName
+      _            -> liftTx <$> GR.mutFldToTx fld
+    return (G.unName $ G.unAlias $ VQ._fAlias fld, fldRespTx)
+
+  -- combines all transactions into a single transaction
+  return $ toSingleTx aliasedTxs
+  where
+    -- A list of aliased transactions for eg
+    -- [("f1", Tx r1), ("f2", Tx r2)]
+    -- are converted into a single transaction as follows
+    -- Tx {"f1": r1, "f2": r2}
+    toSingleTx :: [(Text, LazyRespTx)] -> LazyRespTx
+    toSingleTx aliasedTxs =
+      fmap encJFromAssocList $
+      forM aliasedTxs $ \(al, tx) -> (,) al <$> tx
+
+getMutOp
+  :: (MonadError QErr m)
+  => GCtx
+  -> SQLGenCtx
+  -> UserInfo
+  -> VQ.SelSet
+  -> m LazyRespTx
+getMutOp ctx sqlGenCtx userInfo selSet =
+  runE ctx sqlGenCtx userInfo $ resolveMutSelSet selSet
+
+getSubsOpM
+  :: ( MonadError QErr m
+     , MonadReader r m
+     , Has OpCtxMap r
+     , Has FieldMap r
+     , Has OrdByCtx r
+     , Has SQLGenCtx r
+     , Has UserInfo r
+     )
+  => [G.VariableDefinition]
+  -> VQ.Field
+  -> m (EL.LiveQueryOp, Maybe EL.SubsPlan)
+getSubsOpM varDefs fld =
+  case VQ._fName fld of
+    "__typename" ->
+      throwVE "you cannot create a subscription on '__typename' field"
+    _            -> do
+      astUnresolved <- GR.queryFldToPGAST fld
+      EL.subsOpFromPGAST varDefs (VQ._fAlias fld, astUnresolved)
+
+getSubsOp
+  :: (MonadError QErr m)
+  => GCtx
+  -> SQLGenCtx
+  -> UserInfo
+  -> VQ.Field
+  -> [G.VariableDefinition]
+  -> m (EL.LiveQueryOp, Maybe EL.SubsPlan)
+getSubsOp gCtx sqlGenCtx userInfo fld varDefs =
+  runE gCtx sqlGenCtx userInfo $ getSubsOpM varDefs fld
 
 execRemoteGQ
   :: (MonadIO m, MonadError QErr m)
@@ -169,35 +331,3 @@ execRemoteGQ manager userInfo reqHdrs q rsi opDef = do
                   , "Accept-Language", "Accept-Datetime"
                   , "Cache-Control", "Connection", "DNT"
                   ]
-
-assertSameLocationNodes
-  :: (MonadError QErr m) => [VT.TypeLoc] -> m VT.TypeLoc
-assertSameLocationNodes typeLocs =
-  case Set.toList (Set.fromList typeLocs) of
-    -- this shouldn't happen
-    []    -> return VT.HasuraType
-    [loc] -> return loc
-    _     -> throw400 NotSupported msg
-  where
-    msg = "cannot mix top level fields from two different graphql servers"
-
--- TODO: we should fix this function asap
--- as this will fail when there is a fragment at the top level
-getTopLevelNodes :: G.TypedOperationDefinition -> [G.Name]
-getTopLevelNodes opDef =
-  mapMaybe f $ G._todSelectionSet opDef
-  where
-    f = \case
-      G.SelectionField fld        -> Just $ G._fName fld
-      G.SelectionFragmentSpread _ -> Nothing
-      G.SelectionInlineFragment _ -> Nothing
-
-gatherTypeLocs :: GCtx -> [G.Name] -> [VT.TypeLoc]
-gatherTypeLocs gCtx nodes =
-  catMaybes $ flip map nodes $ \node ->
-    VT._fiLoc <$> Map.lookup node schemaNodes
-  where
-    schemaNodes =
-      let qr = VT._otiFields $ _gQueryRoot gCtx
-          mr = VT._otiFields <$> _gMutRoot gCtx
-      in maybe qr (Map.union qr) mr
