+++ server/src-lib/Hasura/GraphQL/Validate.hs
@@ -6,6 +6,12 @@ module Hasura.GraphQL.Validate
   , QueryParts (..)
   , getQueryParts
   , getAnnVarVals
+
+  , VarPGTypes
+  , AnnPGVarVals
+  , getAnnPGVarVals
+  , Field(..)
+  , SelSet
   ) where
 
 import           Data.Has
@@ -22,6 +28,9 @@ import           Hasura.GraphQL.Validate.Field
 import           Hasura.GraphQL.Validate.InputValue
 import           Hasura.GraphQL.Validate.Types
 import           Hasura.RQL.Types
+
+import           Hasura.SQL.Types (PGColType)
+import           Hasura.SQL.Value (PGColValue, parsePGValue)
 
 data QueryParts
   = QueryParts
@@ -103,6 +112,34 @@ getAnnVarVals varDefsL inpVals = do
 
 showVars :: (Functor f, Foldable f) => f G.Variable -> Text
 showVars = showNames . fmap G.unVariable
+
+type VarPGTypes = Map.HashMap G.Variable PGColType
+type AnnPGVarVals = Map.HashMap G.Variable PGColValue
+
+-- this is in similar spirit to getAnnVarVals, however
+-- here it is much simpler and can get rid of typemap requirement
+-- combine the two if possible
+getAnnPGVarVals
+  :: (MonadError QErr m)
+  => VarPGTypes
+  -> Maybe VariableValues
+  -> m AnnPGVarVals
+getAnnPGVarVals varTypes varValsM =
+  flip Map.traverseWithKey varTypes $ \varName varType -> do
+    let unexpectedVars = filter
+          (not . (`Map.member` varTypes)) $ Map.keys varVals
+    unless (null unexpectedVars) $
+      throwVE $ "unexpected variables in variableValues: " <>
+      showVars unexpectedVars
+    varVal <- onNothing (Map.lookup varName varVals) $
+      throwVE $ "expecting a value for non-nullable variable: " <>
+      showVars [varName] <>
+      -- TODO: we don't have the graphql type
+      -- " of type: " <> T.pack (show varType) <>
+      " in variableValues"
+    runAesonParser (parsePGValue varType) varVal
+  where
+    varVals = fromMaybe Map.empty varValsM
 
 validateFrag
   :: (MonadError QErr m, MonadReader r m, Has TypeMap r)
